requirements.txt
# Core dependencies
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.65.0

# API and HTTP
requests>=2.31.0
openai>=1.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
